CSS fixes for self-containedness and minor improvements
* All style rules prefixed by .que.coderunner or
  body#page-question-type-coderunner as applicable. This should ensure
  that styles from this question type do not have unintended
  consequecnes in other places.

* Remove unneeded spacer <p>. If necessary this should be done with
  margins, but it looks OK without to me.

* Improve overflow on .coderunner-examples and .coderunner-test-results
so we only get a horizontal scroll-bar.

* Move overflow scroll from div.outcome to .coderunner-test-results
which appears to work better with the other CSS changes. (Without this,
in the OU theme, you get an unwanted vertical scroll bar.)

* In questions, add a wrapper with max-width 100% around Ace instances.
This works better on smaller screens / narrow browser windows. I could
not easily make this work in editing forms, since then max-withd 100%
takes as its reference something that is too wide.

* Nicer padding (in my eyes) for .coderunner-test-results

* Move a few styles that seem to have moved away from their logical
places in relation to the comments.