Fix missing_prototype test
This test is failing when we try to run it, as the textarea that the first scenario tries to use is hidden (its been replaced by the ACE editor).  This fix sets the UI plugin to "none" for the question, which is the same method used by the second scenario.